Remove APIs entirely
