+++ dashboard/app/controllers/report_abuse_controller.rb
@@ -65,37 +65,12 @@ class ReportAbuseController < ApplicationController
       end
 
       if params[:channel_id].present?
-        channels_path = "/v3/channels/#{params[:channel_id]}/abuse"
-        assets_path = "/v3/assets/#{params[:channel_id]}/"
-        files_path = "/v3/files/#{params[:channel_id]}/"
+        channel_id = params[:channel_id]
 
-        _, _, body = ChannelsApi.call(
-          'REQUEST_METHOD' => 'POST',
-          'PATH_INFO' => channels_path,
-          'REQUEST_PATH' => channels_path,
-          'HTTP_COOKIE' => request.env['HTTP_COOKIE'],
-          'rack.input' => StringIO.new
-        )
+        abuse_score = update_channel_abuse_score(channel_id)
 
-        abuse_score = JSON.parse(body[0])["abuse_score"]
-
-        FilesApi.call(
-          'REQUEST_METHOD' => 'PATCH',
-          'PATH_INFO' => assets_path,
-          'REQUEST_PATH' => assets_path,
-          'QUERY_STRING' => "abuse_score=#{abuse_score}",
-          'HTTP_COOKIE' => request.env['HTTP_COOKIE'],
-          'rack.input' => StringIO.new
-        )
-
-        FilesApi.call(
-          'REQUEST_METHOD' => 'PATCH',
-          'PATH_INFO' => files_path,
-          'REQUEST_PATH' => files_path,
-          'QUERY_STRING' => "abuse_score=#{abuse_score}",
-          'HTTP_COOKIE' => request.env['HTTP_COOKIE'],
-          'rack.input' => StringIO.new
-        )
+        update_file_abuse_score('assets', channel_id, abuse_score)
+        update_file_abuse_score('files', channel_id, abuse_score)
       end
     end
 
@@ -108,5 +83,84 @@ class ReportAbuseController < ApplicationController
       email: current_user&.email,
       age: current_user&.age,
     }
+  end
+
+  # MOVED ABUSE API STUFF
+
+  # POST /v3/channels/:channel_id/abuse
+  # Increment an abuse score.
+  def update_channel_abuse_score(channel_id)
+    # Reports of abuse from verified teachers are more reliable than reports
+    # from students so we increase the abuse score enough to block the project
+    # with only one report from a verified teacher.
+    #
+    # Temporarily ignore anonymous reports and only allow verified teachers
+    # and signed in users to report.
+    restrict_reporting_to_verified_users = DCDO.get('restrict-abuse-reporting-to-verified', true)
+    amount =
+      if current_user&.verified_teacher?
+        20
+      elsif current_user && !restrict_reporting_to_verified_users
+        10
+      else
+        0
+      end
+    begin
+      value = Projects.new(get_storage_id).increment_abuse(channel_id, amount)
+    rescue ArgumentError, OpenSSL::Cipher::CipherError
+      raise ActionController::BadRequest.new, "Bad channel_id"
+    end
+    value
+  end
+
+  # PATCH /v3/(animations|assets|sources|files|libraries)/:channel_id?abuse_score=<abuse_score>
+  # Update all assets for the given channelId to have the provided abuse score
+  def update_file_abuse_score(endpoint, encrypted_channel_id, abuse_score)
+    return if abuse_score.nil?
+
+    buckets = get_bucket_impl(endpoint).new
+
+    begin
+      files = buckets.list(encrypted_channel_id)
+    rescue ArgumentError, OpenSSL::Cipher::CipherError
+      raise ActionController::BadRequest.new, "Bad channel_id"
+    end
+    files.each do |file|
+      break unless can_update_abuse_score?(endpoint, encrypted_channel_id, file[:filename], abuse_score)
+      buckets.replace_abuse_score(encrypted_channel_id, file[:filename], abuse_score)
+    end
+
+    abuse_score
+  end
+
+  private
+
+  def project_validator?
+    return false unless current_user
+    return true if current_user.permission?(UserPermission::PROJECT_VALIDATOR)
+    false
+  end
+
+  def get_bucket_impl(endpoint)
+    case endpoint
+    when 'animations'
+      AnimationBucket
+    when 'assets'
+      AssetBucket
+    when 'files'
+      FileBucket
+    when 'sources'
+      SourceBucket
+    when 'libraries'
+      LibraryBucket
+    else
+      raise ActionController::RoutingError, 'Not Found'
+    end
+  end
+
+  def can_update_abuse_score?(endpoint, encrypted_channel_id, filename, new_score)
+    return true if project_validator? || new_score.nil?
+
+    get_bucket_impl(endpoint).new.get_abuse_score(encrypted_channel_id, filename) <= new_score.to_i
   end
 end
